Add files via upload
We have 2 genomes. The first genome has 3 genes: g1, g2, g2. The second genome also has 3 genes: G1, G2, G3.  The user enters the values ​​of the genes, the type of transformation and on which gene he wants the transformation.

Transformations can be:

z - the 1st and 2nd genes of replacing the values ​​of the genes located at the n position.

p - In the 1st genome, the gene at position n takes on the value of the gene at position n of the 2nd genome.

d - In the 1st genome, the value of the gene at n position is reduced by the value of the gene at n position of the 2nd genome.

After the user performs the transformation, the strength of the two genomes is compared. Genome strength is calculated as the sum of all genes of a particular genome. At the end, the program prints the genes of the stronger genome and its strength.